+++ plugin/src/adapter50/kotlin/org/jlleitschuh/gradle/ktlint/worker/KtLintInvocation50.kt
@@ -1,0 +1,45 @@
+package org.jlleitschuh.gradle.ktlint.worker
+
+import com.pinterest.ktlint.rule.engine.api.Code
+import com.pinterest.ktlint.rule.engine.api.KtLintRuleEngine
+import com.pinterest.ktlint.rule.engine.api.LintError
+import com.pinterest.ktlint.ruleset.standard.StandardRuleSetProvider
+import java.io.File
+
+class KtLintInvocation50(
+    private val engine: KtLintRuleEngine
+) : KtLintInvocation {
+    companion object Factory : KtLintInvocationFactory {
+        fun initialize(): KtLintInvocation {
+            val engine = KtLintRuleEngine(
+                ruleProviders = StandardRuleSetProvider().getRuleProviders()
+            )
+            return KtLintInvocation50(engine)
+        }
+    }
+
+    override fun invokeLint(file: File): LintErrorResult {
+        val errors = mutableListOf<Pair<SerializableLintError, Boolean>>()
+        engine.lint(Code.fromFile(file)) { le: LintError ->
+            errors.add(le.toSerializable() to false)
+        }
+        return LintErrorResult(file, errors)
+    }
+
+    override fun invokeFormat(file: File): Pair<String, LintErrorResult> {
+        val errors = mutableListOf<Pair<SerializableLintError, Boolean>>()
+        val newCode =
+            engine.format(Code.fromFile(file)) { le, boolean ->
+                errors.add(le.toSerializable() to boolean)
+            }
+        return newCode to LintErrorResult(file, errors)
+    }
+
+    override fun trimMemory() {
+        engine.trimMemory()
+    }
+}
+
+internal fun LintError.toSerializable(): SerializableLintError {
+    return SerializableLintError(line, col, ruleId.value, detail, canBeAutoCorrected)
+}